requirements.txt
streamlit==1.35.0
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.19.0
kaleido==0.2.1
langchain==0.1.16
langchain-groq==0.1.3
langchain-community==0.0.38
groq==0.4.2
python-dotenv==1.0.1
openpyxl==3.1.2
reportlab==4.1.0
pillow==10.2.0
scikit-learn==1.4.0
scipy==1.12.0
numpy==1.24.3
watchdog==4.0.0